# docs/swagger.yaml
definitions:
...
  dto.TestEmailRequest:
    properties:
      to_email:
        type: string
    required:
    - to_email
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  setting_handler.GetSettingsByKeysReq:
    properties:
      keys:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - keys
    type: object
...
  theme.ThemeFieldCondition:
    properties:
      field:
        description: 依赖的字段名
        type: string
      operator:
        description: '操作符: eq, neq, contains, gt, lt'
        type: string
      value:
        description: 比较值
    type: object
  theme.ThemeFieldValidation:
    properties:
      max:
        description: 最大值（数字类型）
        type: number
      maxLength:
        description: 最大长度
        type: integer
      message:
        description: 验证失败提示
        type: string
      min:
        description: 最小值（数字类型）
        type: number
      minLength:
        description: 最小长度
        type: integer
      pattern:
        description: 正则表达式
        type: string
    type: object
...
  theme.ThemeSettingField:
    properties:
      condition:
        allOf:
        - $ref: '#/definitions/theme.ThemeFieldCondition'
        description: 显示条件（依赖其他字段）
      default:
        description: 默认值
      description:
        description: 字段描述
        type: string
      label:
        description: 显示标签
        type: string
      name:
        description: 字段名称（唯一标识）
        type: string
      options:
        description: 选项（用于 select、radio 类型）
        items:
          $ref: '#/definitions/theme.ThemeSettingOption'
        type: array
      placeholder:
        description: 占位提示
        type: string
      required:
        description: 是否必填
        type: boolean
      type:
        description: '字段类型: text, textarea, number, select, color, switch, image,
          code'
        type: string
      validation:
        allOf:
        - $ref: '#/definitions/theme.ThemeFieldValidation'
        description: 验证规则
    type: object
  theme.ThemeSettingGroup:
    properties:
      fields:
        description: 配置字段列表
        items:
          $ref: '#/definitions/theme.ThemeSettingField'
        type: array
      group:
        description: 分组标识（如 style, layout, footer）
        type: string
      label:
        description: 分组显示名称
        type: string
    type: object
  theme.ThemeSettingOption:
    properties:
      label:
        description: 选项显示名称
        type: string
      value:
        description: 选项值
    type: object
...
info:
  contact: {}
paths:
...
  /proxy/download:
    get:
      description: 代理下载外部资源（主要用于图片）
      parameters:
      - description: 目标URL
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 文件内容
          schema:
            type: file
        "400":
          description: 参数错误
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: 代理失败
          schema:
            properties:
              error:
                type: string
            type: object
        "502":
          description: 目标服务器错误
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 代理下载
      tags:
      - 代理服务
...
  /settings/get-by-keys:
    post:
      consumes:
      - application/json
      description: 根据键名列表批量获取配置项（管理员可获取所有配置，普通用户只能获取公开配置）
      parameters:
      - description: 配置键名列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/setting_handler.GetSettingsByKeysReq'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量获取配置
      tags:
      - 站点设置
  /settings/test-email:
    post:
      consumes:
      - application/json
      description: 根据当前配置发送一封测试邮件到指定地址，用于验证邮件服务是否可用。
      parameters:
      - description: 测试邮件请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TestEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功发送
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 邮件发送失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 发送测试邮件
      tags:
      - 设置管理
  /settings/update:
    post:
      consumes:
      - application/json
      description: 批量更新站点配置项（需要管理员权限）
      parameters:
      - description: 配置项键值对
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量更新配置
      tags:
      - 站点设置
...
  /theme/settings:
    get:
      description: 获取指定主题的配置字段定义（用于后台生成配置表单）
      parameters:
      - description: 主题名称
        in: query
        name: theme_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/theme.ThemeSettingGroup'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取主题配置定义
      tags:
      - 主题配置
...
swagger: "2.0"